+++ changelog.md
@@ -3,6 +3,7 @@
 - Cleanup cache naming
 - Remove attr prefix
 - Remove max_age from most method signatures
+- Simplify data cache
 
 # Version 2023.9.7 (2023-09-29)